caculate mutiple division modify file
